refactor: split embedding model from collection and visualization
